add all of the controlers, add all of the dtos, add validator and  extensions, add all the of the services, add all stripe interfaces, add all json configurations, add businessInfo, add all specifications, add amazon secret configuration options, add all migrations,

+++ API/Dtos/LocationDto.cs
@@ -11,7 +11,7 @@ namespace API.Dtos
 
         [Required]
         [MaxLength(4)]
-        public int[] Bbox { get; set; }
+        public float[] Bbox { get; set; }
 
         [Required]
         public string MapBoxId { get; set; }
